conversation_bot/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
requests
loguru
joblib
certifi
python-dotenv
python-multipart
openai[embeddings]
aiohttp
fastapi-utils
envyaml
tokenizers
pydub
pinecone-client
openai
llama-index
httpx
sqlalchemy
databases
faiss-cpu
langchain
langchain-community
langchain-core
numpy
